Move some navigations sagas to the main process
